# api_server/routes/swagger-ui-dist/swagger.json
{
  "swagger": "2.0",
...
  "host": "localhost:8080",
  "basePath": "/API",
  "schemes": [
    "http"
  ],
  "paths": {
    "/Facts": {
      "get": {
        "tags": [
          "Facts"
        ],
        "summary": "Get all facts from the database",
        "description": "Get request to retrieve all facts from the database",
        "operationId": "",
        "produces": "application/json",
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/GET Fact"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Facts"
        ],
        "summary": "Adds a new facts to the database",
        "description": "Post request to add a new fact to the database",
        "operationId": "",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": "application/json",
        "parameters": [
          {
            "name": "statement",
            "in": "formData",
            "description": "Fact that is to be added",
            "required": true,
            "type": "string"
          },
          {
            "name": "popularity",
            "in": "formData",
            "description": "Popularity rating",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/POST Fact"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/Facts/{FactId}": {
      "get": {
        "tags": [
          "Facts"
        ],
        "summary": "Gets a specific fact, based off its ID, from the database",
        "description": "Get request to fetch a fact from the database",
        "operationId": "",
        "produces": "application/json",
        "parameters": [
          {
            "name": "FactId",
            "in": "path",
            "description": "ID of fact to be fetched",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GET Specific Fact"
            }
          },
          "404": {
            "schema": {
              "$ref": "#/definitions/Not Found"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Facts"
        ],
        "summary": "Removes a fact, based off of it's ID, from the database",
        "description": "Delete request to remove a fact from the database",
        "operationId": "",
        "produces": "application/json",
        "parameters": [
          {
            "name": "FactId",
            "in": "path",
            "description": "ID of source to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DELETE Specific Fact"
            }
          },
          "404": {
            "schema": {
              "$ref": "#/definitions/Not Found"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/Moderators": {
      "get": {
        "tags": [
          "Moderators"
        ],
        "summary": "Get all system moderators from the database",
        "description": "Get request to retrieve system moderators from the database",
        "operationId": "",
        "produces": "application/json",
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/GET Moderator"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Moderators"
        ],
        "summary": "Adds a new moderator to the database",
        "description": "Post request to add a new moderator to the database",
        "operationId": "",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": "application/json",
        "parameters": [
          {
            "name": "emailAddress",
            "in": "formData",
            "description": "Moderator's Email Address",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Moderator's password",
            "required": true,
            "type": "string"
          },
          {
            "name": "fName",
            "in": "formData",
            "description": "Moderator's Firstname",
            "required": true,
            "type": "string"
          },
          {
            "name": "lName",
            "in": "formData",
            "description": "Moderator's Lastname",
            "required": true,
            "type": "string"
          },
          {
            "name": "phoneNumber",
            "in": "formData",
            "description": "Moderator's cell phone number",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/POST Moderator"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/Moderators/{EmailAddress}": {
      "get": {
        "tags": [
          "Moderators"
        ],
        "summary": "Get a specific moderator source from the database",
        "description": "Get request to retrieve a specific moderator from the database",
        "operationId": "",
        "produces": "application/json",
        "parameters": [
          {
            "name": "EmailAddress",
            "in": "path",
            "description": "Email address of moderator",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/GET Specific Moderator"
            }
          },
          "404": {
            "schema": {
              "$ref": "#/definitions/GET Specific Moderator 404"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Moderators"
        ],
        "summary": "Removes a specific moderator, based off of their email address, from the database",
        "description": "Delete request to remove a specific moderator from the database",
        "operationId": "",
        "produces": "application/json",
        "parameters": [
          {
            "name": "EmailAddress",
            "in": "path",
            "description": "Email address of moderator to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DELETE Specific Moderator"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Moderators"
        ],
        "summary": "Update moderators details",
        "description": "Put request to update a moderators details",
        "operationId": "",
        "produces": "application/json",
        "parameters": [
          {
            "name": "EmailAddress",
            "in": "path",
            "description": "Email Address of source to be updated",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "New password for the moderator",
            "required": false,
            "type": "string"
          },
          {
            "name": "fName",
            "in": "formData",
            "description": "Update the moderators firstname",
            "required": false,
            "type": "string"
          },
          {
            "name": "lName",
            "in": "formData",
            "description": "Update the moderators lastname",
            "required": false,
            "type": "string"
          },
          {
            "name": "phoneNumber",
            "in": "formData",
            "description": "Update the moderators phone number",
            "required": false,
            "type": "string"
          },
          {
            "name": "authenticationLevel",
            "in": "formData",
            "description": "Update the moderators authentication level",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/PUT Specific Moderator"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "properties": {
        "response": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Server Error"
            },
            "success": {
              "type": "boolean",
              "example": false
            }
          }
        }
      }
    },
...
  }
}